support w3m images
w3m images are a hack which renders on top of the terminal's drawable,
which didn't work in st because when using double buffering, the front
buffer (on which w3m draws its images) is ignored, and st draws only
on the back buffer, which is then copied to the front buffer.

There's a patch to make it work at the FAQ already, but that patch
canceles double-buffering, which can have negative side effects on
some cases such as flickering.

This patch achieves the same goal but instead of canceling the double
buffer it first copies the front buffer to the back buffer.

This has the same issues as the FAQ patch in that the cursor line is
deleted at the image (because st renders always full lines), but
otherwise it's simpler and does keeps double buffering.